requirements.txt
# Pinned versions for production stability
requests>=2.31.0
supabase>=1.0.3
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
azure-functions

# Added dependency to load environment variables from .env file
python-dotenv>=1.0.0
pandas==2.2.2
